+++ src/main/java/com/eodegano/cocobackend/dto/PoiInfoItemDto.java
@@ -1,0 +1,11 @@
+package com.eodegano.cocobackend.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class PoiInfoItemDto {
+    private String infoname;
+    private String infotext;
+}