+++ src/main/java/com/eodegano/cocobackend/service/PoiDetailService.java
@@ -1,0 +1,7 @@
+package com.eodegano.cocobackend.service;
+
+import com.eodegano.cocobackend.dto.PoiDetailResponseDto;
+
+public interface PoiDetailService {
+    PoiDetailResponseDto getPoiDetail(Long contentId);
+}